refactor(devti): split MCP tool details into separate component #371
- Create new McpToolDetailPanel component for tool details
- Remove tool card and details dialog from McpPreviewEditor
- Update McpPreviewEditor to use McpToolDetailPanel
