+++ core/src/main/kotlin/cc/unitmesh/devti/mcp/editor/McpPreviewEditor.kt
@@ -26,7 +26,6 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
-import kotlinx.serialization.json.JsonElement
 import java.awt.BorderLayout
 import java.awt.Dimension
 import java.awt.FlowLayout
@@ -73,10 +72,6 @@ open class McpPreviewEditor(
     private val primaryBlue = JBColor(0x3B82F6, 0x589DF6) // Equivalent to Tailwind blue-500
     private val textGray = JBColor(0x6B7280, 0x9DA0A8)    // Equivalent to Tailwind gray-500
     
-    // Constants for UI sizing
-    private val MAX_TOOL_CARD_HEIGHT = 180
-    private val TOOL_CARD_WIDTH = 300
-
     init {
         createUI()
         loadTools()
@@ -104,7 +99,6 @@ open class McpPreviewEditor(
             }
         }
     }
-
 
     private fun createUI() {
         val headerPanel = panel {
@@ -235,171 +229,14 @@ open class McpPreviewEditor(
         } else {
             allTools.forEach { (serverName, tools) ->
                 tools.forEach { tool ->
-                    val toolCard = createToolCard(serverName, tool)
-                    toolsContainer.add(toolCard)
+                    val panel = McpToolDetailPanel(project, serverName, tool)
+                    toolsContainer.add(panel)
                 }
             }
         }
         
         toolsContainer.revalidate()
         toolsContainer.repaint()
-    }
-
-    private fun createToolCard(serverName: String, tool: Tool): JPanel {
-        val card = JPanel(BorderLayout(0, 8)).apply {
-            background = UIUtil.getPanelBackground()
-            border = CompoundBorder(
-                BorderFactory.createLineBorder(borderColor),
-                JBUI.Borders.empty(16)
-            )
-            // Set preferred width and maximum height
-            preferredSize = Dimension(TOOL_CARD_WIDTH, MAX_TOOL_CARD_HEIGHT)
-            maximumSize = Dimension(Integer.MAX_VALUE, MAX_TOOL_CARD_HEIGHT)
-        }
-
-        // Card header with icon placeholder and title
-        val headerPanel = JPanel(BorderLayout(8, 4)).apply {
-            background = UIUtil.getPanelBackground()
-        }
-
-        // Icon placeholder
-        val iconPlaceholder = JPanel().apply {
-            preferredSize = Dimension(20, 20)
-            background = UIUtil.getPanelBackground()
-            border = BorderFactory.createLineBorder(primaryBlue, 1)
-        }
-
-        val iconWrapper = JPanel(FlowLayout(FlowLayout.LEFT, 0, 0)).apply {
-            background = UIUtil.getPanelBackground()
-            add(iconPlaceholder)
-        }
-
-        val titleLabel = JBLabel(tool.name).apply {
-            font = JBUI.Fonts.label(16.0f).asBold()
-        }
-
-        val titleWrapper = JPanel(BorderLayout()).apply {
-            background = UIUtil.getPanelBackground()
-            add(titleLabel, BorderLayout.CENTER)
-        }
-
-        val titleRow = JPanel(BorderLayout(8, 0)).apply {
-            background = UIUtil.getPanelBackground()
-            add(iconWrapper, BorderLayout.WEST)
-            add(titleWrapper, BorderLayout.CENTER)
-        }
-
-        // Make description scrollable if it's too long
-        val descriptionText = tool.description ?: "No description available"
-        val descLabel = JBLabel("<html><body style='width: ${TOOL_CARD_WIDTH - 50}px'>$descriptionText (from $serverName)</body></html>").apply {
-            font = JBUI.Fonts.label(14.0f)
-            foreground = textGray
-        }
-        
-        val descScrollPane = JBScrollPane(descLabel).apply {
-            border = BorderFactory.createEmptyBorder()
-            verticalScrollBar.unitIncrement = 8
-            background = UIUtil.getPanelBackground()
-            preferredSize = Dimension(TOOL_CARD_WIDTH - 32, 70) // Control description height
-        }
-
-        headerPanel.add(titleRow, BorderLayout.NORTH)
-        headerPanel.add(descScrollPane, BorderLayout.CENTER)
-
-        // Card footer with button
-        val footerPanel = JPanel(BorderLayout()).apply {
-            background = UIUtil.getPanelBackground()
-        }
-
-        val detailsButton = JButton("Details").apply {
-            font = JBUI.Fonts.label(14.0f)
-            isFocusPainted = false
-            addActionListener { showToolDetails(serverName, tool) }
-        }
-
-        footerPanel.add(detailsButton, BorderLayout.CENTER)
-
-        card.add(headerPanel, BorderLayout.CENTER)
-        card.add(footerPanel, BorderLayout.SOUTH)
-
-        return card
-    }
-
-    private fun showToolDetails(serverName: String, tool: Tool) {
-        val dialog = object : DialogWrapper(project) {
-            init {
-                title = "Tool Details"
-                init()
-            }
-
-            override fun createCenterPanel(): JComponent {
-                return panel {
-                    row {
-                        label(tool.name).applyToComponent {
-                            font = JBUI.Fonts.label(18.0f).asBold()
-                        }
-                    }
-                    row {
-                        label("From server: $serverName").applyToComponent {
-                            font = JBUI.Fonts.label(14.0f)
-                            foreground = textGray
-                        }
-                    }
-                    row {
-                        label(tool.description ?: "No description available").applyToComponent {
-                            font = JBUI.Fonts.label(14.0f)
-                        }
-                    }
-
-                    group("Parameters") {
-                        tool.inputSchema.properties?.forEach { param: Map.Entry<String, JsonElement> ->
-                            row {
-                                label("${param.key}")
-                                    .applyToComponent {
-                                        font = JBUI.Fonts.label(14.0f)
-                                    }
-                            }
-                            row {
-                                label(param.value.toString())
-                                    .applyToComponent {
-                                        font = JBUI.Fonts.label(12.0f)
-                                        foreground = textGray
-                                    }
-                            }
-                        }
-                    }
-                }.withPreferredSize(400, 200)
-            }
-
-            override fun createSouthPanel(): JComponent {
-                val panel = JPanel(FlowLayout(FlowLayout.RIGHT)).apply {
-                    background = UIUtil.getPanelBackground()
-                }
-
-                val executeButton = JButton("Execute").apply {
-                    font = JBUI.Fonts.label(14.0f)
-                    addActionListener {
-                        executeTool(serverName, tool)
-                        close(OK_EXIT_CODE)
-                    }
-                }
-
-                panel.add(executeButton)
-                return panel
-            }
-        }
-
-        dialog.show()
-    }
-
-    private fun executeTool(serverName: String, tool: Tool) {
-        val result = mcpServerManager.execute(project, tool, "{}")
-        JOptionPane.showMessageDialog(
-            component,
-            result,
-            "Tool Execution Result",
-            JOptionPane.INFORMATION_MESSAGE
-        )
     }
 
     private fun showConfigDialog() {
@@ -433,7 +270,6 @@ open class McpPreviewEditor(
                                 addChangeListener {
                                     val value = temperatureSlider.value / 10.0
                                     config.temperature = value
-//                                    this.text = "Temperature: ${String.format("%.1f", value)}"
                                 }
                             })
                         }
@@ -453,7 +289,6 @@ open class McpPreviewEditor(
                                 addChangeListener {
                                     val value = tokensSlider.value
                                     config.maxTokens = value
-//                                    component.text = "Max Tokens: $value"
                                 }
                             })
                         }
